+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import Header from "../../components/header";
+import Header from "../components/organisms/Header";
 
 export default function Home() {
   return (
